Add files via upload
A bitmap font for localizing pixel art games.

LanaPixel contains approximately 19400 characters, with support for most modern European languages (Latin, Greek, Cyrillic), Turkish (Latin), and Korean (Hangul), good support for Japanese and Simplified Chinese, and decent support for Traditional Chinese. The Latin, Greek, and Cyrillic characters have kerning.

This font is designed to be displayed at 11px. As this is a pixel font, it is recommended to only scale by integer amounts (1x or 11px, 2x or 22px, 3x or 33px...) to avoid distortion.